server/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
Pillow>=10.1
psycopg2-binary>=2.9
python-dotenv>=1.0
celery[redis]>=5.3
whitenoise>=6.6
xhtml2pdf>=0.2.11
requests>=2.31
cryptography>=42.0
channels>=4.0
channels-redis>=4.2
pywebpush>=1.14
google-generativeai>=0.3.0
django-celery-beat>=2.5
